Rename files to dbreg.
